Move utils to helper folder
