+++ src/components/index.tsx
@@ -20,8 +20,10 @@ export { Header, HeaderContent } from './Header';
 export { Layout } from './Layout';
 export { Library, LibraryProps, LibraryContent } from './Library';
 export { Link } from './Link';
+export { PageWithPosts } from './PageWithPosts';
 export { PostContainer, PostContent } from './PostContainer';
 export { Section, SectionWithTitle } from './Section';
+export { SEO } from './seo';
 export {
   SocialLink,
   SocialLinkWithMaterial,
