+++ src/helper/string.ts
