+++ src/pages/404.tsx
@@ -25,7 +25,7 @@ import { footerContent } from '../content/FooterContent';
 import { headerContent } from '../content/HeaderContent';
 
 import { Layout, Section, MailingListSubscription } from '../components';
-import { PAGE } from '../theme/utils/constants';
+import { PAGE } from '../helper';
 
 import Img from '../assets/images/404.png';
