+++ src/pages/index.tsx
@@ -23,7 +23,7 @@ import { DataProps } from '../hooks/use-site-metadata';
 import { Landing, Blog, News, About, Libraries } from '../sections';
 
 import { Layout, MailingListSubscription } from '../components';
-import { PAGE } from '../theme/utils/constants';
+import { PAGE } from '../helper';
 
 import { footerContent } from '../content/FooterContent';
 import { headerContent } from '../content/HeaderContent';
