+++ src/pages/model-generation-application-conditions.tsx
@@ -30,7 +30,7 @@ import { footerContent } from '../content/FooterContent';
 import { headerContent } from '../content/HeaderContent';
 
 import { Section, Layout } from '../components';
-import { PAGE } from '../theme/utils/constants';
+import { PAGE } from '../helper';
 
 const ModelGenerationApplicationConditionPage = (): JSX.Element => (
   <Layout footerContent={footerContent} headerContent={headerContent}>
