+++ src/sections/home/Libraries.tsx
@@ -18,7 +18,7 @@ import React from 'react';
 import { Container, Grid } from '@mui/material';
 
 import { Library, SectionWithTitle } from '../../components';
-import { SECTION } from '../../theme/utils/constants';
+import { SECTION } from '../../helper/constants';
 
 import { libraryContent } from '../../content/LibraryContent';
